composition: split into 3 files
